+++ Changelog.md
@@ -72,6 +72,7 @@
 * Change `DocxStructure#parse_paragraph_style_xml` to instance method
 * Redone parsing `Shape`
 * Remove deprecated `GridSpan#type` and `GridSpan#count_of_merged_cells`
+* Remove unused `ParagraphMargins#round`
   
 ### Fixes
 * Fix crash on empty coordinates list of chart